機材情報のページでのロジックをカスタムHookで分離
